+++ package.json
@@ -1732,13 +1732,15 @@
 		"build": "rm -rf dist && tsc",
 		"vscode:prepublish": "webpack --mode production",
 		"webpack": "webpack --mode development",
-		"webpack-dev": "webpack --mode development --watch"
+		"webpack-dev": "webpack --mode development --watch",
+		"typings": "npx -p typescript tsc ./src/global.d.ts ./src/extension.ts --declaration --allowJs --emitDeclarationOnly --outDir ../vscode-ibmi-types --esModuleInterop && cd ../vscode-ibmi-types && rm -rf types/filesystems types/schemas types/views types/webviews types/languages"
 	},
 	"devDependencies": {
 		"@types/glob": "^7.1.3",
 		"@types/node": "^12.11.7",
 		"@types/source-map-support": "^0.5.6",
 		"@types/vscode": "^1.66.0",
+		"dets": "^0.12.2",
 		"eslint": "^7.19.0",
 		"glob": "^7.1.6",
 		"semver": "^7.3.5",
