+++ src/filesystems/ifs.js
@@ -1,7 +1,7 @@
 
 const util = require(`util`);
 const vscode = require(`vscode`);
-let {instance} = require(`../Instance`);
+let {instance} = require(`../instantiate`);
 
 module.exports = class qsysFs {
   constructor() {
